I actually implemented switch, not add lol

+++ src/commands/switch.js
@@ -4,44 +4,43 @@ import path from 'path';
 import uuid from 'uuid';
 import fs from 'fs';
 import { promisify } from 'util';
-// import addCommand from '../logger';
+import { addCommand } from '../logger';
 import resolve from '../utils/resolve';
 import fsResolver from '../utils/fsResolver';
 
+import { getConfigDirectoryPath, configDirectoryExists } from '../utils/file';
+
 // import type {Resolver, ChangeIntent} from '../utils/resolve';
 
-const backupDir = path.resolve(process.env.HOME, './.sweesh/backups');
-// const pluginsDir = path.resolve(process.env.HOME, './.sweesh/plugins');
+const backupDir = getConfigDirectoryPath('backups');
+const pluginsDir = getConfigDirectoryPath('plugins');
 
 const copyFile = promisify(fs.copyFile);
 const writeFile = promisify(fs.writeFile);
+const readdir = promisify(fs.readdir);
 
-export const command = 'add <app|document-name> [document-path]';
+export const command = 'switch <app|document-name> [document-path]';
 
-export const describe = 'Add a file to sweesh-cli servers';
+export const describe = 'Switch a local file to one stored on server';
 
 export const builder = {
 };
 
 export async function handler(argv: any) {
-    const resolvers = [];
+    if (!configDirectoryExists()) {
+        throw Error('Make sure you login first');
+    }
     let changeIntent = null;
     if (argv.documentPath) {
         changeIntent = await resolve([fsResolver], argv.documentPath);
     }
-
     else {
-        // TODO: get all the resolvers from the resolvers directory
-        // TODO: call resolve with all resolvers and the file name
+        const pluginFolders = await readdir(pluginsDir);
+        const resolvers = pluginFolders.map(plugin => path.join(pluginsDir, plugin)).map(require);
         changeIntent = await resolve(resolvers, argv.app);
     }
-    // const rootDir = changeIntent.path.split('/').slice(0, -1).join('/') + '/';
-    // TODO: execute CHANGE intent
     const backupFile = path.join(backupDir, uuid());
-    // TODO: create dir if it doesn't exist
     await copyFile(changeIntent.path, backupFile);
-    // TODO: fetch actual file from server
     writeFile(changeIntent.path, 'Test config file', 'utf8');
-    // TODO: write change log to logger
-    // addCommand('add', changeIntent.path, backupFile);
+    addCommand('add', backupFile, changeIntent.path);
 };
